RR Single Pipeline
Flipped the resolve relationship pipeline from one that was nested into a straight pipeline that splits and merges each entry. splitting each entry in order to resolve each relationship independently in the pipeline, and the last step of the pipeline is to merge them back together, merging the data before it gets saved.

Missing all the save functionality, testing to see if it works.