Plus Diffing & Saving
The relationship resolving pipeline now updates related and current object data, and depending on if they were updated ornot, saves them.

This type of logic would be more helpful if we were getting more granular data dumps. But since everthing is coming in as one large dump, containing every object, every time. It will be more sane to do the same thing on our end, and remove all of the relationships, and programmatically entered objects, and allow the system to do its thing.