RR!
Relationships have been resolved for Courses.

In doing so, found out that the course key was not appropriate, since coursename is not unique to all courses. They can appear in different departments, with different terms. Oh my. So the key will now be the course unique id, and the course name in order to maintain some readbility in webhook relationship UI.