Revisiting Reverse Relationships
The strategy employeed currently will over write new data with old data. The reverse relationships should be concatenated, but they are stored in an array. Firebase doesn't do arrays, particularly ones that are meant to be unique. So relationships will be resolved, and then their revere will be resolved after.