Reverse Relationships Realigned
Got Events set up to test out the new approach on reverse relationships. Stash reverse data to be stored in the Firebase, under a eduSync/reverseRelationships node. This is used to store the reverse keys and values as data flows through the relationship pipeline.

Once the relationship pipeline is complete, the reverseRelationship pipeline is initiated, and the eduSync/reverseRelationship nodes are traversed and used to update the webhook data.