One-Off reverse
Turns out the value of a relationship is the name of the content type twice. Where multiple content types, the value is the name of the content type, and the internally used firebase key. Since there the internal firebase key for one off content types is just the name of the content type, I guessed relationship values were just the name. Verified, and fixed with this commit.

Also getting Courses up to speed with the one-off relationships it has to graduate studies and foundation studies.