New Relationships
Added libaeralartsdepartments relationships to courses & employees.
Added employees relationship to courses.
Updated the way the department mapping works. Since the structure of the departments is that Liberal Arts Departments are seperate from the other Academic departments, they get their own map structure for comparing name representations from different sources. Doing this made refactoring the way those map functions are defined make sense, a thing that has made more sense as more functions that do the same thing were being made.
Updated the way data for relationship resolution was being captured. Related data, which is an entire content type's data in webhook, was being captured for every object that came through the relationship resolution pipeline. This related data is actually only used to for referece. To know which keys need to be stored so that they can be related to. Since the data does not change, keeping a cache of it is just fine.