Report First cut
Any sources that make it through the pipeline also get written out to the eduSyncReport node, which has keys of content types, and values of the date they were last synced.

Once the values for the current sources are updated, all values are fetched, written to HTML, and pushed to S3.

edu-data-sync-report.s3-website-us-east-1.amazonaws.com