Report Complete
Changed the package used to upload to S3, the previous one did not allow me to define the content-type, so the file would always be seen as an application & be downloaded rather than viewed in browser.

Added a 404.html, manually put it on S3 since it won't be changing.

Added some space between source entries in the report. Should take a moment and give the stylesheet some love.