Python Scheduler
In my searching, I did not find a node scheduler that looked solid. Memory leaks & all kinds of talk about how I should be using cron if I want a task to be run repeatedly over time. But I have experience with the APSScheduler in python, and it has worked well for this kind of thing. So this is the start of having a python scheduler, and testing of it works to run node processes.