Event within date range
Added a new field to the Event content type, within_date_range. This will be set based on sync, every hour, and will be populated with a value of true or false based on the time of sync, and if at that time is between the start and the end date of the event.